fixed terminal command to disown process from terminal
Before, the "start-command" for spotify-notify only put the script in the background. However, as the terminal session was ended (i.e. the window was closed), the spotify-notify-script was terminated, too. This behaviour is now fixed by a "disown" command at the end. It unbinds the spotify-notify.py command from the terminal session, so that closing the terminal does not end the spotify-notify process anymore